all I'm doing for now probably worth adding the concept of proxying into the ecs itself
